Split up dynamic keyboard location settings.
We now have the choice of:
- built in keyboards, whose path is fixed in appdata, and is adaptive to
whether point source is mouse or eye tracker.
- user-specified folder
- single file

This means that DynamicKeyboardSelector and related classes now take a
path as input rather than rely on settings, and the settings UI has been edited
to support the new options.
Old settings files should be upgraded to the new scheme.